codegen(java): C-ABI-correct struct layout (bool/enum/tag)
Java JNA struct field layout was misaligned vs the C ABI in three ways:

1. `bool` mapped to Java `boolean` which JNA marshals as 4-byte Win32
   BOOL by default. Switched to `byte` so AzTitlebarButtons (4 bools)
   is 4 bytes both sides instead of 16.

2. Unit-enum struct fields used the Java `enum AzX` class. JNA can't
   size heap-object enum fields, so any struct containing one (e.g.
   AzAppConfig.log_level = AzAppLogLevel) failed Structure layout
   validation with "Native size unknown". Switched unit-enum struct
   fields to `int` (matches C `enum X` = sizeof(int)).

3. Tagged-union variant `tag` was `int` (4 bytes) but the C ABI uses
   `#[repr(C, u8)]` (1 byte). On small-aligned variants (Option<u8>,
   Option<ColorU>) this turned a 5-byte struct into 8; AzSystemColors
   {21 × Option<ColorU>} was 168 bytes vs C 105. Tag now `byte`,
   helper constructors cast `(byte) AzX_Tag.Foo.value`.

All sampled types now match C: AzAppConfig 1640/1640, AzWindowCreateOptions
1344/1344, AzFullWindowState 1264/1264, AzSystemStyle 1288/1288,
AzSystemColors 105/105, AzOptionColorU 5/5.

Java full-GUI hello-world is still blocked by a separate JNA limit on
JDK 21+: the auto-generated Proxy class for `AzulNative` (1700+ method
signatures) exceeds JVM's 64KB method-bytecode limit when building
`<clinit>`. Fix is a follow-up codegen split of AzulNative into
per-namespace interfaces.

+++ doc/src/codegen/v2/lang_java/types.rs
@@ -248,7 +248,10 @@ fn emit_monomorphized_alias_files(
                             variant_struct
                         ));
                         b.indent();
-                        b.line(&format!("public int tag; // {}_Tag.{}", name, v.name));
+                        // C-ABI `#[repr(C, u8)]` tag is 1 byte. Using Java `byte` (8-bit)
+// matches that; `int` would be 4 bytes and shift every payload
+// offset on small-aligned variants.
+b.line(&format!("public byte tag; // {}_Tag.{}", name, v.name));
                         let mut field_names: Vec<String> = vec!["\"tag\"".to_string()];
                         emit_monomorphized_payload(b, v, ir, &mut field_names);
                         emit_field_order_override(b, &field_names);
@@ -448,7 +451,10 @@ fn emit_tagged_union_files(
                     variant_struct
                 ));
                 b.indent();
-                b.line(&format!("public int tag; // {}_Tag.{}", name, v.name));
+                // C-ABI `#[repr(C, u8)]` tag is 1 byte. Using Java `byte` (8-bit)
+// matches that; `int` would be 4 bytes and shift every payload
+// offset on small-aligned variants.
+b.line(&format!("public byte tag; // {}_Tag.{}", name, v.name));
 
                 let mut field_names: Vec<String> = vec!["\"tag\"".to_string()];
                 match &v.kind {
